start processPacket;  SystemHeader processing to terminate after length
The system header previously checked what the next bit was. This however ended inserting an additional byte into the bit_buffer. In an example file, this ended up consume the next 0x00 in a system start code.

Instead since we know the length of the header in bytes, just check that instead.