fix processPacket; introduce readBits31515
The main issue was that the stream id is part of the start code which is consumed outside of the processPacket call. Set it in the main marker reading loop. Add asserts that the bit_buffer is empty after processing a start code. Additionally we can have an 8 bit pattern stuffed which will be consmed.