add mb_motion_vector VLCs
Since the motion vectors can be negative, but are restricted to -16 we'll just bitCast them. Since comptime ints don't have a defined bit length, first cast them to i8 and then bitCast to u8.

This sort of messes up generateLookupTables check since we're using 255 as a sentinel value for no entry. Technically the check is still correct but we won't detect overwrites of 255.