add isSliceCode to start_codes enum
There's 175 possible slice start codes, and it feels annoying to generate the enum with all those values.

Instead add a method to check if the code is a slice code by checking if the int value falls between 1 and 175.

Then in the marker loop, under the else branch, check the is slice. This is sort  of annoying because it kind of defeats the purpose of an exhaustive switch.