demuxer and video decoder split
Previously the architecture had a single decoder loop; the loop would process ALL start code types. So in this case it mixed demuxer/system level packets with the video packets.

When the single loop encountered any packet it would immediately start processing that packet. In this case after detecting the video_stream_0 system header it would automatically start calling the video decoding functions when it encountered video system start codes (like start slice). This is a problem because a single packet only contains a fragment of a frame. So the lowest level call would need to detect that it ran out of the current packet, and either process it or kick it back out to the demuxer level.

Instead we sort of want to the video/audio decoders to request packets from the demuxer. So the video decoder requests that the demuxer send data for some time interval and that fills its buffer.

Split the start code processing into two loops, one that just decodes system level packets and send it to the appropriate decoder ie in this case the video only. Right now the demuxer automatically processes my entire video file since it's small. and then we process the video packets in a giant buffer. In the future we'll only send the necessary packets to decode the frame.

Split start codes into VideoStartCodes and SystemStartCodes.