have a single function return start code
findNextStartCode now takes a type, which should be one of VideoStartCodes or SystemStartCodes (not enforced) and then tries to find and return the next start code of that type. Panics otherwise.

This is done to reduce the amount of code for each enum. There's probably a better way to do this, interface? etc.